Chk for set-env.ps1 in the 1st "up", more licenses

+++ windows/build/os/centos-wm-install-helper/commands.ps1
@@ -2,19 +2,25 @@
 $PSThisScriptRoot = Split-Path -Parent -Path $MyInvocation.MyCommand.Definition
 
 Write-Output "Script folder: $PSThisScriptRoot"
-Invoke-Expression -Command ". $PSThisScriptRoot\..\..\..\common\config\set-env.ps1"
 
-$command=$args[0]
+if( Test-Path $PSThisScriptRoot\..\..\..\common\config\set-env.ps1 -PathType Leaf){
+    Invoke-Expression -Command ". $PSThisScriptRoot\..\..\..\common\config\set-env.ps1"
 
-switch  -exact -casesensitive  ($command) {
-    "up"      { docker-compose up;      break }
-    "down"    { docker-compose down;    break }
-    "destroy" { docker-compose down -v; break }
-    "stop"    { docker-compose stop;    break }
-    "start"   { docker-compose start;   break }
-    "shell"   { docker exec -ti wm-install-helper-base bash; break }
-    Default   { Write-Host "Unknown command $command";       break }
+    $command=$args[0]
+
+    switch  -exact -casesensitive  ($command) {
+        "up"      { docker-compose up;      break }
+        "down"    { docker-compose down;    break }
+        "destroy" { docker-compose down -v; break }
+        "stop"    { docker-compose stop;    break }
+        "start"   { docker-compose start;   break }
+        "shell"   { docker exec -ti wm-install-helper-base bash; break }
+        Default   { Write-Host "Unknown command $command";       break }
+    }  
+}else{
+    Write-Host "Project set-env.ps1 file does not exist, set up the project first!"
 }
+
 
 Write-Host "Stopping for debug"
 pause
